Made classes modules, started preparation fix cohesion issue
