+++ start.py
@@ -19,7 +19,7 @@ if __name__ == "__main__":
         os.environ[item] = conf[item]
 
     # Note that the state is not saved when running this development file.
-    import DLAPI
-    DLAPI.main()
+    import dlapi.DLAPI
+    dlapi.DLAPI.main()
     input()
